Move ReloadableConfigInjectorProvider to rune-testing for downstream reuse
Generalized: no baked-in config resource — useConfig(...) serves a classpath
resource as rune-config.yml and resetConfig() restores the default lookup, so
downstream tests (e.g. the BSP's rosetta/updateConfig tests) can exercise live
config reloads with the same utility.
